Add user deletion
Add delete button to settings. Set userselection to NoResize. Automatically reload the combobox after deleting a user. Automatically reload the ComboBox on import of a user. Handle exceptions for certian file operations which might occur if there is a lock on the file. Produce relevant error messages and return in such instances.